backend/requirements.txt
fastapi
uvicorn
pydantic
torch
torchvision
fastapi
uvicorn
web3
py-solc-x
numpy
